requirements.txt
psycopg2-binary
SQLAlchemy
sqlparse
pandas==2.1.4